FIX kassenbuch time comparisons: discard all microseconds
From now on, kassenbuch discards all microseconds when filtering for a specific time range.

This may create a race-condition if a booking is made at xxx.100 seconds and then `kassenbuch show` is called at xxx.200 sec, because the booking was made after xxx.000 sec
and the `snapshot_time` mechanism filters out the booking from the "future".