FIX FAUCardPayment: Rework user abortion, termination and finish_log
In summary, the code that handles the end of a payment is now improved, but still does the same as before. Probably a few bugs were fixed on the way.

Remove the possibility to force-terminate the thread.
Program behavior of QThread terminate() is not well-defined and may leave the program/memory in an undefined state, so we must not allow this.

Remove confusing dialog box about user abortion.

Remove confusing set_finish_log. Make clear that finish_log may only be used outside FAUCardThread.

Rework thread closing.

Remove some unused functions.